+++ kitty/colors.c
@@ -170,7 +170,7 @@ patch_color_profiles(PyObject *module UNUSED, PyObject *args) {
 }
         S(foreground, default_fg); S(background, default_bg); S(cursor, cursor_color);
         S(selection_foreground, highlight_fg); S(selection_background, highlight_bg);
-        S(cursor_text_color, cursor_text_color);
+        S(cursor_text_color, cursor_text_color); S(visual_bell_color, visual_bell_color);
 #undef SI
 #undef S
     Py_RETURN_NONE;
@@ -234,7 +234,7 @@ as_dict(ColorProfile *self, PyObject *args UNUSED) {
     }}
     D(default_fg, foreground); D(default_bg, background);
     D(cursor_color, cursor); D(cursor_text_color, cursor_text); D(highlight_fg, selection_foreground);
-    D(highlight_bg, selection_background);
+    D(highlight_bg, selection_background); D(visual_bell_color, visual_bell_color);
 
 #undef D
     return ans;
@@ -295,13 +295,13 @@ set_color(ColorProfile *self, PyObject *args) {
 static PyObject*
 set_configured_colors(ColorProfile *self, PyObject *args) {
 #define set_configured_colors_doc "Set the configured colors"
-    unsigned int default_fg, default_bg, cursor_color, cursor_text_color, highlight_fg, highlight_bg;
-    if (!PyArg_ParseTuple(args, "II|IIII", &default_fg, &default_bg,
-        &cursor_color, &cursor_text_color, &highlight_fg, &highlight_bg)) return NULL;
+    unsigned int default_fg, default_bg, cursor_color, cursor_text_color, highlight_fg, highlight_bg, visual_bell_color;
+    if (!PyArg_ParseTuple(args, "II|IIIII", &default_fg, &default_bg,
+        &cursor_color, &cursor_text_color, &highlight_fg, &highlight_bg,  &visual_bell_color)) return NULL;
 #define S(which) \
     self->configured.which.rgb = which & 0xffffff; \
     self->configured.which.type = (which & 0xff000000) ? COLOR_IS_RGB : COLOR_IS_SPECIAL;
-    S(default_fg); S(default_bg); S(cursor_color); S(cursor_text_color); S(highlight_fg); S(highlight_bg);
+    S(default_fg); S(default_bg); S(cursor_color); S(cursor_text_color); S(highlight_fg); S(highlight_bg); S(visual_bell_color);
 #undef S
     self->dirty = true;
     Py_RETURN_NONE;
@@ -412,6 +412,7 @@ CGETSET(cursor_color)
 CGETSET(cursor_text_color)
 CGETSET(highlight_fg)
 CGETSET(highlight_bg)
+CGETSET(visual_bell_color)
 #undef CGETSET
 
 static PyGetSetDef cp_getsetters[] = {
@@ -421,6 +422,7 @@ static PyGetSetDef cp_getsetters[] = {
     GETSET(cursor_text_color)
     GETSET(highlight_fg)
     GETSET(highlight_bg)
+    GETSET(visual_bell_color)
     {NULL}  /* Sentinel */
 };